unet-sar/requirements.txt
# Requirements for U-Net SAR Segmentation
# Pure NumPy implementation with optional GPU acceleration

# Core dependencies
numpy>=1.19.0

# Optional: GPU acceleration via CuPy
# Install based on your CUDA version:
# For CUDA 11.x: pip install cupy-cuda11x>=11.0.0
# For CUDA 12.x: pip install cupy-cuda12x>=12.0.0
# If CuPy is not installed, the code will automatically fall back to NumPy (CPU)
# cupy-cuda11x>=11.0.0  # Uncomment for CUDA 11.x
# cupy-cuda12x>=12.0.0  # Uncomment for CUDA 12.x

# Optional: For image loading and manipulation
pillow>=8.0.0

# Optional: For advanced image resizing (COCO dataset)
scipy>=1.7.0

# Optional: For ONNX model export (MATLAB integration)
onnx>=1.10.0
onnx-simplifier>=0.3.0

# Optional: For centralized YAML configuration
pyyaml>=6.0

# Note: np_mlp_autograd is included in parent directory
# No installation needed for np_mlp_autograd
